SNOW-206907 GCS Downscope Token
Description
- Implement update to work with downscoped accesstoken instead of using
presigned url. To use this new feature, SNOW-206907 must be reviewed and merged
on the server side first.
- The upload and download proccess is unchanged, XML api URL is built and used
with the access token set in header when access token is available.
- `get_file_header()` now works with gcs with parameter turn on
- Add more tests!!!
- Move file existence for overwrite check into the look, with old behavior, error
will cause it to bypass the existence check and just try overwrite, which may success.
This would also retry getting the file header.

Testing
- All put/get test for gcs test with parameter GCS_USE_DOWNSCOPED_CREDENTIAL turn
both on and off
- Test modify to work with OVERWRITE
- Additional error code test.

+++ src/snowflake/connector/remote_storage_util.py
@@ -79,23 +79,6 @@ class SnowflakeRemoteStorageUtil(object):
 
         util_class = SnowflakeRemoteStorageUtil.getForStorageType(
             meta['stage_info']['locationType'])
-        if not meta.get('overwrite'):
-            file_header = util_class.get_file_header(
-                meta, meta['dst_file_name'])
-            if meta['result_status'] == ResultStatus.RENEW_TOKEN:
-                # need renew token
-                return
-            elif meta['result_status'] == ResultStatus.RENEW_PRESIGNED_URL:
-                return
-            elif file_header and \
-                    meta['result_status'] == ResultStatus.UPLOADED:
-                logger.debug(
-                    'file already exists location="%s", file_name="%s"',
-                    meta['stage_info']['location'],
-                    meta['dst_file_name'])
-                meta['dst_file_size'] = 0
-                meta['result_status'] = ResultStatus.SKIPPED
-                return
 
         logger.debug('putting a file: %s, %s',
                      meta['stage_info']['location'], meta['dst_file_name'])
@@ -104,12 +87,27 @@ class SnowflakeRemoteStorageUtil(object):
         last_err = None
         max_retry = DEFAULT_MAX_RETRY
         for retry in range(max_retry):
-            util_class.upload_file(
-                data_file,
-                meta,
-                encryption_metadata,
-                max_concurrency
-            )
+            if not meta.get('overwrite'):
+                file_header = util_class.get_file_header(
+                    meta, meta['dst_file_name'])
+
+                if file_header and \
+                        meta['result_status'] == ResultStatus.UPLOADED:
+                    logger.debug(
+                        'file already exists location="%s", file_name="%s"',
+                        meta['stage_info']['location'],
+                        meta['dst_file_name'])
+                    meta['dst_file_size'] = 0
+                    meta['result_status'] = ResultStatus.SKIPPED
+                    return
+
+            if meta.get('overwrite') or meta.get('result_status') == ResultStatus.NOT_FOUND_FILE:
+                util_class.upload_file(
+                    data_file,
+                    meta,
+                    encryption_metadata,
+                    max_concurrency
+                )
 
             if (meta['result_status'] == ResultStatus.UPLOADED):
                 return
@@ -194,7 +192,7 @@ class SnowflakeRemoteStorageUtil(object):
                     # preserve the idea of getting metadata in the first place.
                     # One example of this is the utils that use presigned url
                     # for upload/download and not the storage client library.
-                    if meta['stage_location_type'] == 'GCS':
+                    if meta.get('presigned_url', None):
                         file_header = util_class.get_file_header(meta, meta[
                             'src_file_name'])
